Safari-609.1.11.1/LayoutTests/fast/dom/HTMLAnchorElement/anchor-file-blob-download-then-revoke-expected.txt

CONSOLE MESSAGE: line 37: PASS: URL was revoked
Download started.
Downloading URL with suggested filename "foo.txt"
Download completed.
The suggested filename above should be "foo.txt" and the download should succeed.

Memory backed blob URL
